refactor + dockerized FastAPI, streamlit apps
