+++ api/config/api_config.yaml
@@ -1,0 +1,4 @@
+is_cuda: false
+img_resize: 256
+beam_size: 10
+model_config: !include ./model.yaml
