+++ api/tests/config.py
@@ -4,7 +4,7 @@ REQUEST_ROUTE = '/inference'
 
 SUCCESS_CODE = 200
 
-IMAGE_PATH = './demo/demo_img1.jpg'
+IMAGE_PATH = '../app/demo/demo_img1.jpg'
 
 SENTENCE_CLASSES = [0, 1, 2]
 EMOJI_CLASSES = [0, 1]
