+++ app/config/app_config.yaml
@@ -1,0 +1,3 @@
+demo_image: ./demo/demo_img1.jpg
+page_icon: https://raw.githubusercontent.com/riven314/TweakStory/master/assets/icon.png
+img_display: 512
