--- app/src/models.py
@@ -1,295 +1,0 @@
-import os
-import json
-
-import torch
-from torch import nn
-import torch.nn.functional as F
-import torchvision
-
-from src.dropouts import RNNDropout, InputDropout, WeightDropout, EmbeddingDropout
-
-
-def get_encoder_decoder(cfg):
-    with open(cfg.word_map_file, 'r') as f:
-        word_map = json.load(f)
-
-    encoder = Encoder()
-    encoder.fine_tune(cfg.fine_tune_encoder)
-
-    dropout_dict = cfg.regularized_decoder
-    embed_p, weight_p  = dropout_dict['embed_p'], dropout_dict['weight_p']
-    input_p, output_p = dropout_dict['input_p'], dropout_dict['output_p']
-    decoder = RegularizedDecoderWithAttention(
-        attention_dim = cfg.attention_dim,
-        embed_dim = cfg.emb_dim, decoder_dim = cfg.decoder_dim,
-        style_dim = cfg.style_dim, vocab_size = len(word_map),
-        embed_p = embed_p, weight_p = weight_p, 
-        input_p = input_p, output_p = output_p
-        )
-
-    if cfg.checkpoint_file is not None:
-        assert os.path.isfile(cfg.checkpoint_file)
-        checkpoint = torch.load(cfg.checkpoint_file, map_location = lambda storage, loc: storage)
-        encoder.load_state_dict(checkpoint['encoder'])
-        decoder.load_state_dict(checkpoint['decoder'])
-        print(f'loaded checkpoint in CPU: {cfg.checkpoint_file}')
-    return encoder, decoder
-
-
-class Encoder(nn.Module):
-    """
-    Encoder.
-    """
-
-    def __init__(self, encoded_image_size=14):
-        super(Encoder, self).__init__()
-        self.enc_image_size = encoded_image_size
-
-        resnet = torchvision.models.resnet101(pretrained=True)  # pretrained ImageNet ResNet-101
-
-        # Remove linear and pool layers (since we're not doing classification)
-        modules = list(resnet.children())[:-2]
-        self.resnet = nn.Sequential(*modules)
-
-        # Resize image to fixed size to allow input images of variable size
-        self.adaptive_pool = nn.AdaptiveAvgPool2d((encoded_image_size, encoded_image_size))
-
-        self.fine_tune()
-
-    def forward(self, images):
-        """
-        Forward propagation.
-
-        :param images: images, a tensor of dimensions (batch_size, 3, image_size, image_size)
-        :return: encoded images
-        """
-        out = self.resnet(images)  # (batch_size, 2048, image_size/32, image_size/32)
-        out = self.adaptive_pool(out)  # (batch_size, 2048, encoded_image_size, encoded_image_size)
-        out = out.permute(0, 2, 3, 1)  # (batch_size, encoded_image_size, encoded_image_size, 2048)
-        return out
-
-    def fine_tune(self, fine_tune=True):
-        """
-        Allow or prevent the computation of gradients for convolutional blocks 2 through 4 of the encoder.
-
-        :param fine_tune: Allow?
-        """
-        for p in self.resnet.parameters():
-            p.requires_grad = False
-        # If fine-tuning, only fine-tune convolutional blocks 2 through 4
-        for c in list(self.resnet.children())[5:]:
-            for p in c.parameters():
-                p.requires_grad = fine_tune
-
-
-class Attention(nn.Module):
-    """
-    Attention Network.
-    """
-
-    def __init__(self, encoder_dim, decoder_dim, attention_dim):
-        """
-        :param encoder_dim: feature size of encoded images
-        :param decoder_dim: size of decoder's RNN
-        :param attention_dim: size of the attention network
-        """
-        super(Attention, self).__init__()
-        self.encoder_att = nn.Linear(encoder_dim, attention_dim)  # linear layer to transform encoded image
-        self.decoder_att = nn.Linear(decoder_dim, attention_dim)  # linear layer to transform decoder's output
-        self.full_att = nn.Linear(attention_dim, 1)  # linear layer to calculate values to be softmax-ed
-        self.relu = nn.ReLU()
-        self.softmax = nn.Softmax(dim=1)  # softmax layer to calculate weights
-
-    def forward(self, encoder_out, decoder_hidden):
-        """
-        Forward propagation.
-
-        :param encoder_out: encoded images, a tensor of dimension (batch_size, num_pixels, encoder_dim)
-        :param decoder_hidden: previous decoder output, a tensor of dimension (batch_size, decoder_dim)
-        :return: attention weighted encoding, weights
-        """
-        att1 = self.encoder_att(encoder_out)  # (batch_size, num_pixels, attention_dim)
-        att2 = self.decoder_att(decoder_hidden)  # (batch_size, attention_dim)
-
-        att = self.full_att(self.relu(att1 + att2.unsqueeze(1))).squeeze(2)  # (batch_size, num_pixels)
-        alpha = self.softmax(att)  # (batch_size, num_pixels)
-        attention_weighted_encoding = (encoder_out * alpha.unsqueeze(2)).sum(dim=1)  # (batch_size, encoder_dim)
-
-        return attention_weighted_encoding, alpha
-
-
-class RegularizedDecoderWithAttention(nn.Module):
-    """ Decoder without regularization """
-    def __init__(self, attention_dim, embed_dim, decoder_dim, vocab_size, 
-                 encoder_dim = 2048, style_dim = 128, 
-                 embed_p = 0.1, weight_p = 0.5, input_p = 0.6, output_p = 0.3):
-        """
-        :param attention_dim: size of attention network
-        :param embed_dim: embedding size
-        :param decoder_dim: size of decoder's RNN
-        :param vocab_size: size of vocabulary
-        :param encoder_dim: feature size of encoded images
-        :param hidden_p: hidden state dropout 
-        :param embed_p: embedding droppout
-        :param weight_p: hidden-to-hidden weight matrix dropout
-        """
-        super(RegularizedDecoderWithAttention, self).__init__()
-
-        self.encoder_dim = encoder_dim
-        self.attention_dim = attention_dim
-        self.embed_dim = embed_dim
-        self.decoder_dim = decoder_dim
-        self.vocab_size = vocab_size
-        self.length_class_embed_dim = style_dim
-
-        self.attention = Attention(encoder_dim, decoder_dim, attention_dim)  # attention network
-
-        decode_step = nn.LSTMCell(embed_dim + encoder_dim + style_dim + style_dim, decoder_dim, bias=True)
-        self.decode_step_dp = WeightDropout(decode_step, weight_p, layer_names = ['weight_hh'])
-        self.init_h = nn.Linear(encoder_dim, decoder_dim)  # linear layer to find initial hidden state of LSTMCell
-        self.init_c = nn.Linear(encoder_dim, decoder_dim)  # linear layer to find initial cell state of LSTMCell
-        self.f_beta = nn.Linear(decoder_dim, encoder_dim)  # linear layer to create a sigmoid-activated gate
-        self.sigmoid = nn.Sigmoid()
-
-        # embedding layers
-        self.embedding = nn.Embedding(vocab_size, embed_dim)  # embedding layer
-        self.embedding_dp = EmbeddingDropout(self.embedding, embed_p)
-        self.length_class_embedding = nn.Embedding(3, style_dim)
-        self.is_emoji_embedding = nn.Embedding(2, style_dim)
-
-        self.fc = nn.Linear(decoder_dim, vocab_size)  # linear layer to find scores over vocabulary
-        self.init_weights()  # initialize some layers with the uniform distribution
-
-        # dropout layers
-        self.input_dp = RNNDropout(p = input_p)
-        self.output_dp = RNNDropout(p = output_p)
-
-
-    def init_weights(self):
-        """
-        Initializes some parameters with values from the uniform distribution, for easier convergence.
-        """
-        self._init_embedding()
-        self._init_fc()
-
-    def _init_embedding(self):
-        self.embedding.weight.data.uniform_(-0.1, 0.1)
-        self.length_class_embedding.weight.data.uniform_(-0.1, 0.1)
-        self.is_emoji_embedding.weight.data.uniform_(-0.1, 0.1)
-
-    def _init_fc(self):
-        self.fc.bias.data.fill_(0)
-        self.fc.weight.data.uniform_(-0.1, 0.1)
-
-    def load_pretrained_embeddings(self, embeddings):
-        """
-        Loads embedding layer with pre-trained embeddings.
-
-        :param embeddings: pre-trained embeddings
-        """
-        weight = torch.from_numpy(embeddings).float()
-        self.embedding.weight = nn.Parameter(weight)
-        self.embedding.weight.requires_grad = False
-
-    def fine_tune_embeddings(self, fine_tune = True):
-        """
-        Allow fine-tuning of embedding layer? (Only makes sense to not-allow if using pre-trained embeddings).
-
-        :param fine_tune: Allow?
-        """
-        for p in self.embedding.parameters():
-            p.requires_grad = fine_tune
-
-    def init_hidden_state(self, encoder_out):
-        """
-        Creates the initial hidden and cell states for the decoder's LSTM based on the encoded images.
-
-        :param encoder_out: encoded images, a tensor of dimension (batch_size, num_pixels, encoder_dim)
-        :return: hidden state, cell state
-        """
-        mean_encoder_out = encoder_out.mean(dim = 1)
-        h = self.init_h(mean_encoder_out)  # (batch_size, decoder_dim)
-        c = self.init_c(mean_encoder_out)
-        return h, c
-
-    def forward(self, encoder_out, encoded_captions, caption_lengths, length_class, is_emoji):
-        """
-        Forward propagation.
-
-        :param encoder_out: encoded images, a tensor of dimension (batch_size, enc_image_size, enc_image_size, encoder_dim)
-        :param encoded_captions: encoded captions, a tensor of dimension (batch_size, max_caption_length)
-        :param caption_lengths: caption lengths, a tensor of dimension (batch_size, 1)
-        :param length_class: a Long tensor of dim (batch_size, 1)
-        :param is_emoji: a Long tensor of dim (batch_size, 1)
-        :return: scores for vocabulary, sorted encoded captions, decode lengths, weights, sort indices
-        """
-        batch_size = encoder_out.size(0)
-        encoder_dim = encoder_out.size(-1)
-        vocab_size = self.vocab_size
-
-        # Flatten image
-        encoder_out = encoder_out.view(batch_size, -1, encoder_dim)  # (batch_size, num_pixels, encoder_dim)
-        num_pixels = encoder_out.size(1)
-
-        # Sort input data by decreasing lengths; why? apparent below
-        caption_lengths, sort_ind = caption_lengths.squeeze(1).sort(dim = 0, descending=True)
-        encoder_out = encoder_out[sort_ind]
-        encoded_captions = encoded_captions[sort_ind]
-        length_class = length_class[sort_ind]
-        is_emoji = is_emoji[sort_ind]
-
-        # Embedding
-        embeddings = self.embedding_dp(encoded_captions)  # (batch_size, max_caption_length, embed_dim)
-
-        # style embedding
-        length_class = length_class.squeeze()
-        is_emoji = is_emoji.squeeze()
-        len_class_embedding = self.length_class_embedding(length_class)
-        is_emoji_embedding = self.is_emoji_embedding(is_emoji)
-
-        # Initialize LSTM state
-        h, c = self.init_hidden_state(encoder_out)  # (batch_size, decoder_dim)
-
-        # We won't decode at the <end> position, since we've finished generating as soon as we generate <end>
-        # So, decoding lengths are actual lengths - 1
-        decode_lengths = (caption_lengths - 1).tolist()
-
-        # Create tensors to hold word predicion scores and alphas
-        predictions = torch.zeros(batch_size, max(decode_lengths), vocab_size).to(encoder_out.device)
-        alphas = torch.zeros(batch_size, max(decode_lengths), num_pixels).to(encoder_out.device)
-
-        # At each time-step, decode by
-        # attention-weighing the encoder's output based on the decoder's previous hidden state output
-        # then generate a new word in the decoder with the previous word and the attention weighted encoding
-
-        for t in range(max(decode_lengths)):
-            
-            batch_size_t = sum([l > t for l in decode_lengths])
-            attention_weighted_encoding, alpha = self.attention(encoder_out[:batch_size_t],
-                                                                h[:batch_size_t])
-            gate = self.sigmoid(self.f_beta(h[:batch_size_t]))  # gating scalar, (batch_size_t, encoder_dim)
-            attention_weighted_encoding = gate * attention_weighted_encoding
-
-            # concat with word embedding, image-attention encoding, style embedding
-            
-            reset_input_mask = True if t == 0 else False
-            style_embeddings = torch.cat([
-                embeddings[:batch_size_t, t, :], 
-                len_class_embedding[:batch_size_t], 
-                is_emoji_embedding[:batch_size_t]
-            ], dim = 1)
-            dp_style_embeddings = self.input_dp(style_embeddings, reset_mask = reset_input_mask)
-            cat_embeddings = torch.cat([dp_style_embeddings, attention_weighted_encoding], dim = 1)
-            
-            reset_decoder_mask = True if t == 0 else False
-            # (batch_size_t, decoder_dim)
-            h, c = self.decode_step_dp(cat_embeddings, 
-                                       (h[:batch_size_t], c[:batch_size_t]), 
-                                       reset_mask = reset_decoder_mask)  
-            
-            reset_output_mask = True if t == 0 else False
-            preds = self.fc(self.output_dp(h, reset_mask = reset_output_mask))  # (batch_size_t, vocab_size)
-            predictions[:batch_size_t, t, :] = preds
-            alphas[:batch_size_t, t, :] = alpha
-
-        return predictions, encoded_captions, decode_lengths, alphas, sort_ind
